# assets/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Swagger CodeCraft",
        "description": "API endpoints documentations for CodeCraft Project"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/organizations": {
            "get": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Returns the first page of organizations from the system",
                "operationId": "getOrganizations",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "organizations response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/organizations/me": {
            "get": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Returns a list of organizations that a user is a member of.",
                "operationId": "getUserOrganizations",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "organizations response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Organization"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/organizations/{id}": {
            "get": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Returns a Organization based on a single ID",
                "operationId": "getOrganizationById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Organization to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "organization response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/Organization"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/organizations/{id}/job_posts": {
            "get": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Returns the job posts of an organization",
                "operationId": "getJobPosts",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Organization to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "job posts response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/JobPost"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/organizations/{id}/members": {
            "get": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Returns the first page of organization members from the system",
                "operationId": "getOrganizationMembers",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Organization",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "members response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Member"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/organizations/create": {
            "post": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Create a new organization",
                "operationId": "createOrganization",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "create organization response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "data": {
                                    "$ref": "#/definitions/Organization"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/organizations/{id}/update": {
            "post": {
                "tags": [
                    "Organizations Management"
                ],
                "description": "Update Organization",
                "operationId": "updateOrganization",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Organization",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "description": "Bearer token for authentication"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update organization response",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "example": "success"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Organization": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "creator": {
                    "$ref": "#/definitions/User"
                },
                "profile_image_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                }
            }
        },
...
        "JobPost": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-02-28T20:20:48.000Z"
                }
            }
        },
...
        "Member": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                },
                "organization_id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "ErrorModel": {
            "type": "object",
            "required": [
                "status",
                "message"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}